\name{jaquith}
\alias{jaquith}
\docType{data}
\title{Sample Business-Adjusted Risk Data}

\description{
  This dataset provides the sample business-adjusted risk from
  \emph{Security Metrics} to illustrate non-financial waterfall charts.
}
\usage{data(jaquith)}
\format{A data frame with 9 rows and 2 columns.
  \tabular{rlll}{
    [,1] \tab factor \tab character \tab Factor label\cr
    [,2] \tab score  \tab numeric   \tab Relative score\cr
  }
}
\details{The dataset represents a sample business-adjusted risk
  calculation.}

\references{
  Andrew Jaquith, \emph{Security Metrics: Replacing Fear, Uncertainty,
  and Doubt} (Boston:  Addison-Wesley Professional, 2007), 170-171.
}

\keyword{datasets}
