% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plotting the waterfall results with ggplot2}
\usage{
plot_waterfall(df)
}
\arguments{
\item{df}{the output from waterfall}
}
\description{
Plotting the waterfall results with ggplot2
}
