% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{pseudotimeprog.foo}
\alias{pseudotimeprog.foo}
\title{Execute waterfall}
\usage{
pseudotimeprog.foo(x, k = 10, x.reverse = F, plot = F, color = NULL,
  ...)
}
\arguments{
\item{x}{your data}

\item{k}{the number of clusters}

\item{x.reverse}{whether or not to reverse the trajectory}

\item{plot}{whether or not to plot the results}

\item{color}{the colours to be used in the plot}

\item{...}{extra parameters for the plot}
}
\description{
Execute waterfall
}
