\name{rasiel}
\alias{rasiel}
\docType{data}
\title{Sample Financial Data}
\description{
  This dataset provides the sample financial data used in \emph{The
    McKinsey Way} to illustrate financial waterfall charts.
}  

\usage{data(rasiel)}
\format{A data frame with 5 rows and 3 columns.
  \tabular{rlll}{
    [,1] \tab label    \tab character \tab Column label\cr
    [,2] \tab value    \tab numeric   \tab Column height\cr
    [,3] \tab subtotal \tab character \tab Group and subtotal labels\cr
  }
}
\details{The dataset represents a profit and loss statement for the
  fictional ACME Widget Corporation for 1998.}

\references{
  Ethan M. Rasiel, \emph{The McKinsey Way: Using the Techniques of the
    World's Top Strategic Consultants to Help You and Your Business}
  (New York:  McGraw-Hill, 1999), 115-116.
}

\keyword{datasets}
