% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\docType{package}
\name{waterfall-package}
\alias{waterfall}
\alias{waterfall-package}
\title{Waterfall Charts}
\description{
Create waterfall or "McKinsey" charts
}
\details{
This package provides support for creating waterfall
charts in R using both traditional base and lattice graphics.
}
\author{
James P. Howard, II <jh@jameshoward.us>
}

