% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{pseudotimeprog.foo}
\alias{pseudotimeprog.foo}
\title{Execute waterfall}
\usage{
pseudotimeprog.foo(x, k = 10, x.reverse = F)
}
\arguments{
\item{x}{your data}

\item{k}{the number of clusters}

\item{x.reverse}{whether or not to reverse the trajectory}
}
\description{
Execute waterfall
}
