\name{waterfall-package}
\alias{waterfall}
\title{Waterfall Charts in R}

\description{Creates waterfall charts in R using traditional and lattice
graphics} 

\details{This package contains a collection of graphical utilities and
  sample data for demonstration.  For a complete list, use
  \code{library(help="waterfall")}}.

\author{James Howard <jh@jameshoward.us>}

\keyword{hplot}
  
